fix(root): extract BaseEntity/SubscribeEntity to common/entities to break cross-module entity cycles
