+++ apps/aurora-core/src/modules/common/entities/base-entity.ts
