+++ apps/aurora-core/src/modules/common/entities/index.ts
@@ -1,0 +1,2 @@
+export { default as BaseEntity } from './base-entity';
+export { default as SubscribeEntity } from './subscribe-entity';
