+++ apps/aurora-core/src/modules/common/entities/subscribe-entity.ts
