+++ apps/aurora-core/src/modules/handlers/base-handler.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from 'uuid';
-import SubscribeEntity from '../root/entities/subscribe-entity';
+import SubscribeEntity from '../common/entities/subscribe-entity';
 import { BeatEvent } from '../events/music-emitter-events';
 import logger from '../../logger';
