+++ apps/aurora-core/src/modules/lights/entities/lights-controller.ts
@@ -2,7 +2,7 @@ import { Entity, OneToMany } from 'typeorm';
 // eslint-disable-next-line import/no-cycle
 import LightsGroup from './lights-group';
 import LightsSwitch from './lights-switch';
-import SubscribeEntity from '../../root/entities/subscribe-entity';
+import SubscribeEntity from '../../common/entities/subscribe-entity';
 
 @Entity()
 export default class LightsController extends SubscribeEntity {
