+++ apps/aurora-core/src/modules/lights/entities/lights-fixture.ts
@@ -1,5 +1,5 @@
 import { AfterLoad, Column } from 'typeorm';
-import BaseEntity from '../../root/entities/base-entity';
+import BaseEntity from '../../common/entities/base-entity';
 import { RgbColor } from '../color-definitions';
 import { DEFAULT_MASTER_DIMMER } from './lights-group-fixture';
 import { jsonTransformer } from '../../../helpers/transformers';
