+++ apps/aurora-core/src/modules/lights/entities/lights-group-moving-head-rgbs.ts
@@ -1,5 +1,4 @@
 import { Column, Entity, JoinColumn, ManyToOne } from 'typeorm';
-import BaseEntity from '../../root/entities/base-entity';
 // eslint-disable-next-line import/no-cycle
 import LightsGroup from './lights-group';
 import LightsMovingHeadRgb from './lights-moving-head-rgb';
