+++ apps/aurora-core/src/modules/lights/entities/lights-group.ts
@@ -7,7 +7,7 @@ import LightsGroupPars from './lights-group-pars';
 import LightsGroupMovingHeadWheels from './lights-group-moving-head-wheels';
 // eslint-disable-next-line import/no-cycle
 import LightsGroupMovingHeadRgbs from './lights-group-moving-head-rgbs';
-import SubscribeEntity from '../../root/entities/subscribe-entity';
+import SubscribeEntity from '../../common/entities/subscribe-entity';
 
 @Entity()
 export default class LightsGroup extends SubscribeEntity {
