+++ apps/aurora-core/src/modules/lights/entities/lights-wheel-channel-value.ts
@@ -1,7 +1,7 @@
 import { Column, JoinColumn, ManyToOne } from 'typeorm';
 // eslint-disable-next-line import/no-cycle -- cross reference
 import LightsMovingHeadWheel from './lights-moving-head-wheel';
-import BaseEntity from '../../root/entities/base-entity';
+import BaseEntity from '../../common/entities/base-entity';
 
 export default class LightsWheelChannelValue<T> extends BaseEntity {
   @ManyToOne(() => LightsMovingHeadWheel)
