+++ apps/aurora-core/src/modules/root/entities/index.ts
@@ -3,6 +3,7 @@ import Audio from './audio';
 import Screen from './screen';
 
 export { default as Audio } from './audio';
+export { BaseEntity, SubscribeEntity } from '../../common/entities';
 export { default as Screen } from './screen';
 
 export const Entities = [Audio, Screen];
