+++ apps/aurora-core/src/modules/root/entities/screen.ts
@@ -1,5 +1,5 @@
 import { Column, Entity } from 'typeorm';
-import SubscribeEntity from './subscribe-entity';
+import SubscribeEntity from '../../common/entities/subscribe-entity';
 
 @Entity()
 export default class Screen extends SubscribeEntity {
