+++ apps/aurora-core/src/modules/timed-events/entities/timed-event.ts
@@ -1,4 +1,4 @@
-import BaseEntity from '../../root/entities/base-entity';
+import BaseEntity from '../../common/entities/base-entity';
 import { Column, Entity } from 'typeorm';
 import EventSpec from '../event-spec';
 import { jsonTransformer } from '../../../helpers/transformers';
